pyproject.toml
[project]
name = "mdata-flow"
version = "0.0.6rc0"
description = ""
authors = [{ name = "Ilia Novikov", email = "nov.ilucha@yandex.ru" }]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "mlflow-skinny[databricks] (>=2.22.0,<2.23.0)",
  "pydantic-settings (>=2.8.1,<3.0.0)",
  "xxhash (>=3.5.0,<4.0.0)",
  "pydantic (>=2.10.6,<3.0.0)",
  "pandas (>=2.2.3,<3.0.0)",
  "scipy (>=1.15.3,<2.0.0)",
]

[project.optional-dependencies]
graphics = [
  "plotly (>=5.10.0,<6)",
  "matplotlib (>=3.10.1,<4.0.0)",
  "seaborn (>=0.13.2,<0.14.0)",
]
reports = ["evidently (>=0.6.5,<0.7.0)", "plotly (>=5.10.0,<6)"]
compress = ["zstandard (>=0.23.0,<0.24.0)"]

[tool.poetry]
packages = [{ include = "mdata_flow", from = "src" }]


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"
pytest = "^8.3.5"
allure-pytest = "^2.14.2"
pandas-stubs = "^2.2.3.250308"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pipx-install]
poetry = "~=2.1.1"
